Enhance Goal service with user-specific operations
Updated the Goal service to include user-specific operations. Added a new function `getById` to fetch goals owned by a specific user. For the `create`, `update`, and `del` functions, now checking if the authenticated user is authorized to perform the operations on the specified goal based on the owner ID.

+++ back-end/services/goalService.js
@@ -8,26 +8,52 @@ async function get() {
     }
 }
 
-async function create(data) {
+async function getById(user) {
     try {
-        return await Goal.create(data);
+        return await Goal.find({ owner: user._id }).lean();
+    } catch (error) {
+        throw error
+    }
+}
+
+async function create(user, data) {
+    try {
+        const goals = await Goal.create({
+            text: data.text,
+            owner: user._id
+        });
+
+        return goals
 
     } catch (error) {
         throw error
     }
 }
 
-async function update(id, data) {
+async function update(id, data, user) {
     try {
-        return await Goal.findByIdAndUpdate(id, data);
+        const goal = await Goal.findById(id);
+
+        if (goal.owner.toString() !== user._id) {
+            throw new Error('User not authorized')
+        }
+
+        return await Goal.findByIdAndUpdate(id, data, user);
     } catch (error) {
         throw error
     }
 }
 
-async function del(id) {
+async function del(id, user) {
     try {
-        return await Goal.findByIdAndDelete(id);
+        const goal = await Goal.findById(id)
+
+        if (goal.owner.toString() !== user._id) {
+            throw new Error('User not authorized')
+        }
+        console.log(id);
+        await Goal.findByIdAndDelete(id);
+
     } catch (error) {
         throw error
     }
@@ -37,5 +63,6 @@ module.exports = {
     get,
     create,
     update,
-    del
+    del,
+    getById
 }
